tek_access/requirements.txt
altgraph==0.10.2
appdirs==1.4.3
asgiref==3.2.3
astroid==1.6.6
backports.functools-lru-cache==1.6.1
bdist-mpkg==0.5.0
cffi==1.14.0
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==2.8
distlib==0.3.0
Django==3.0.3
django-rest-knox==4.1.0
djangorestframework==3.11.0
enum34==1.1.6
filelock==3.0.12
importlib-metadata==1.5.0
importlib-resources==1.0.2
isort==4.3.21
lazy-object-proxy==1.4.3
macholib==1.5.1
mccabe==0.6.1
modulegraph==0.10.4
pathlib2==2.3.5
psycopg2==2.8.4
py2app==0.7.3
pycparser==2.19
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
wrapt==1.11.2
zipp==3.0.0
